refactor(db_maintenance): separate the scripts from the modules they use
The folder mixed three kinds of file — scripts you run, the pure modules
that decide what those scripts write, and the modules' tests — with nothing
but a naming habit to tell them apart. `scripts/` now holds the things you
run, and the top level holds the library and its tests, so the split that
keeps the test suite dependency-free is visible in the layout.

Two defaults resolved from the working directory and would have broken
quietly:

`dotenv` was loaded bare, so it found `.env` only when you happened to run
from `src/db_maintenance`. Anyone who cd'd into the new folder and ran a
script would have got "MONGODB_URL not set" and no clue why. Each script now
points dotenv at `../.env` relative to itself. `.env` has not moved.

`--backup-dir` / `--out` / `--dir` defaulted to `<script>/backups`, which
would have become `scripts/backups` — snapshots taken before this change
would still exist but no restore would look for them. They now resolve to
`src/db_maintenance/backups` as before, still the path `.gitignore` covers.
